Feat: [FN-302] 그룹 권한 추가

+++ src/main/java/flipnote/group/application/service/DeleteGroupService.java
@@ -6,6 +6,8 @@ import org.springframework.transaction.annotation.Transactional;
 import flipnote.group.application.port.in.DeleteGroupUseCase;
 import flipnote.group.application.port.in.command.DeleteGroupCommand;
 import flipnote.group.application.port.out.GroupRepositoryPort;
+import flipnote.group.application.port.out.GroupRoleRepositoryPort;
+import flipnote.group.domain.model.member.GroupMemberRole;
 import lombok.RequiredArgsConstructor;
 
 @Service
@@ -13,10 +15,20 @@ import lombok.RequiredArgsConstructor;
 public class DeleteGroupService implements DeleteGroupUseCase {
 
 	private final GroupRepositoryPort groupRepository;
+	private final GroupRoleRepositoryPort groupRoleRepository;
 
 	@Override
 	@Transactional
 	public void deleteGroup(DeleteGroupCommand cmd) {
+
+		//오너 인지 확인
+		boolean isOwner = groupRoleRepository.checkRole(cmd.userId(), cmd.groupId(), GroupMemberRole.OWNER);
+
+		//오너가 아닐 경우 에러 발생
+		if(!isOwner) {
+			throw new IllegalArgumentException("not owner");
+		}
+
 		groupRepository.delete(cmd.groupId());
 	}
